[sfmData] Rename `imageInfo*` to ̀ImageInfo*`
