+++ src/aliceVision/sfmData/CMakeLists.txt
@@ -8,7 +8,7 @@ set(sfmData_files_headers
   uid.hpp
   colorize.hpp
   exif.hpp
-  imageInfo.hpp
+  ImageInfo.hpp
   exposureSetting.hpp
   Observation.hpp
 )
@@ -20,7 +20,7 @@ set(sfmData_files_sources
   View.cpp
   colorize.cpp
   exif.cpp
-  imageInfo.cpp
+  ImageInfo.cpp
   Observation.cpp
 )
