+++ src/aliceVision/sfmData/ImageInfo.cpp
@@ -4,7 +4,7 @@
 // v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 
-#include "imageInfo.hpp"
+#include "ImageInfo.hpp"
 
 #include <boost/algorithm/string.hpp>
 #include <aliceVision/numeric/gps.hpp>
