+++ src/aliceVision/sfmData/ImageInfo.hpp
