+++ src/aliceVision/sfmData/View.hpp
@@ -14,7 +14,7 @@
 #include <aliceVision/camera/IntrinsicInitMode.hpp>
 #include <aliceVision/lensCorrectionProfile/lcp.hpp>
 
-#include <aliceVision/sfmData/imageInfo.hpp>
+#include <aliceVision/sfmData/ImageInfo.hpp>
 #include <aliceVision/sfmData/exif.hpp>
 #include <aliceVision/sfmData/exposureSetting.hpp>
